+++ src/Network/Network.Test/ScenarioTests/VirtualNetworkGatewayConnectionTests.cs
@@ -80,5 +80,13 @@ namespace Commands.Network.Test.ScenarioTests
         {
             TestRunner.RunTestScript("Test-VirtualNetworkGatewayConnectionVpnDeviceConfigurations");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        [Trait(Category.Owner, NrpTeamAlias.brooklynft_subset1)]
+        public void TestVirtualNetworkGatewayConnectionWithTrafficSelectorCRUD()
+        {
+            TestRunner.RunTestScript("Test-VirtualNetworkGatewayConnectionWithTrafficSelector");
+        }
     }
 }